Docs i18n: make translation prompt language-pluggable
